% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPESIndicatorMatrix.R
\name{getPESIndicator}
\alias{getPESIndicator}
\title{PES - Indicator Matrix}
\usage{
getPESIndicator(seasons)
}
\arguments{
\item{seasons}{}
}
\description{
A function that takes a vector of seasons, where the i-th entry = k indicates that datapoint i is in season k, and forms
the indicator matrix required for the PES method
}
\examples{
seasons <- c(1,2,3,3)
getPESIndicator(seasons)
1 0 0 0
0 1 0 0
0 0 1 1
}

