% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanAbsoluteError.R
\name{MeanAbsoluteError}
\alias{MeanAbsoluteError}
\title{Mean Absolute Error (MAE)}
\usage{
MeanAbsoluteError(ForecastVec, ObservedVec)
}
\arguments{
\item{ForecastVec}{The vector of forecasts}

\item{ObservedVec}{The vector of observations}
}
\description{
Computes MAE of forecasts.
}

