% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanAbsolutePercentageError.R
\name{MeanAbsolutePercentageError}
\alias{MeanAbsolutePercentageError}
\title{Mean Absolute Percentage Error (MAPE)}
\usage{
MeanAbsolutePercentageError(ForecastVec, ObservedVec)
}
\arguments{
\item{ForecastVec}{The vector of forecasts}

\item{ObservedVec}{The vector of observations}
}
\description{
Computes MAPE of forecasts.
}

