% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RootMeanSquaredError.R
\name{RootMeanSquaredError}
\alias{RootMeanSquaredError}
\title{Root Mean Squared Error (RMSE)}
\usage{
RootMeanSquaredError(ForecastVec, ObservedVec)
}
\arguments{
\item{ForecastVec}{The vector of forecasts}

\item{ObservedVec}{The vector of observations}
}
\description{
Computes RMSE of forecasts.
}

